requirements.txt
graph_theory==2022.4.3
psycopg2_binary==2.9.6
pyyaml~=6.0
requests==2.31.0
SQLAlchemy==1.4.50
tableauserverclient==0.25
urllib3==1.26.18
apache-airflow==2.7.1
apache-airflow-providers-airbyte==3.3.0
apache-airflow-providers-tableau==4.2.0
apache-airflow-providers-snowflake==5.1.1
dbt-core==1.6.8
dbt-snowflake==1.6.4
git+https://github.com/carrerapcl/airflow-dbt-snowflake.git@main#egg=airflow_dbt_snowflake # PyPi registrations are currently disabled...
